requirements.txt
pillow
pandas
openpyxl
reportlab
